refactor: extract game-stats component, add some styling to tables
